Update model. Remove deprecated fields. 	modified:   CHANGELOG.md 	modified:   lavalink_rs/model/http.pyi 	modified:   src/model/http.rs 	modified:   src/player_context/context.rs 	modified:   src/player_context/mod.rs 	modified:   src/python/model/http.rs 	modified:   src/python/model/player.rs

+++ src/python/model/player.rs
@@ -108,7 +108,7 @@ impl TremoloVibrato {
 #[pymethods]
 impl Filters {
     getter_setter!(
-        (volume, Option<u16>),
+        (volume, Option<f64>),
         (equalizer, Option<Vec<Equalizer>>),
         (karaoke, Option<Karaoke>),
         (timescale, Option<Timescale>),
